The dock floats, the feed follows your finger
The bottom bar becomes a floating glass pill with its own outline all the way round, deeper and more transparent than the band it replaces, and the reels disc is a lit sphere in the accent's hue rather than a dark puck in a white ring. Every colour comes off --bar/--accent, so it re-tints with the theme instead of carrying a palette of its own.

Swiping between categories now tracks the finger for the whole gesture and either completes or springs back on release. The title strip no longer moves with it: a header that slid with the feed read as loose. The feed declares touch-action: pan-y so the browser hands over the horizontal axis immediately instead of spending the first pixels deciding.

Pastel drops its cream ground for a cool near-white and raises the chroma across every tinted surface; the cards were carrying the colour anyway.

The manifest gains the raster icons Chromium requires before it will offer to install: an SVG alone never satisfies the check, whatever the scheme.

+++ src/components/HeaderAside.tsx
@@ -18,7 +18,7 @@ export function HeaderAside() {
   const [here, setHere] = useState('');
 
   useEffect(() => {
-    setToday(new Date().toLocaleDateString('en-GB', { weekday: 'long', day: 'numeric', month: 'long' }));
+    setToday(new Date().toLocaleDateString('en-GB', { weekday: 'short', day: 'numeric', month: 'long' }));
     try { setHere(localStorage.getItem(KEY) ?? ''); } catch { /* first visit */ }
 
     let live = true;
